First checkins to get Plisgo working correctly with Windows 7 64bit.
git-svn-id: https://svn.code.sf.net/p/plisgo/code@87 f301a074-b28b-4858-84c8-f8b36c4887ce

+++ Plisgo/PlisgoFSFolder.h
@@ -53,7 +53,7 @@ public:
 		FLOAT	= 2
 	};
 
-	ULONG				GetColumnNum() const							{ return m_Columns.size(); }
+	ULONG				GetColumnNum() const							{ return (ULONG)m_Columns.size(); }
 	const std::wstring&	GetColumnHeader(int nIndex) const				{ return m_Columns[nIndex].sHeader; }
 	ColumnAlignment		GetColumnAlignment(int nIndex) const			{ return m_Columns[nIndex].eAlignment; }
 	ColumnType			GetColumnType(int nIndex) const					{ return m_Columns[nIndex].eType; }
